ISSUE-4 Create temporary directory that matches filepath
I get spurious error messages about files not matching the package names
because of the temporary directory that's used to store files for checking.

This changes it to use a temporary directory that's named after the file
path in the repository. This means we should fail that check if the file
is named incorrectly, but not if it's named correctly.